feature: change login page style, add register page
