+++ src/app/(dashboard-routes)/components/Menu/SubMenu.tsx
@@ -1,0 +1,8 @@
+interface SubMenuProps {
+  children: React.ReactNode;
+}
+
+function SubMenu({ children }: SubMenuProps) {
+  return <>{children}</>;
+}
+export default SubMenu;
