+++ src/app/(dashboard-routes)/dashboard/calendar/layout.tsx
@@ -1,0 +1,10 @@
+import { ILayoutDefault } from '@/interfaces/defaults';
+import { Metadata } from 'next';
+
+export const metadata: Metadata = {
+  title: 'Talkative | Turmas'
+};
+
+export default function Layout({ children }: ILayoutDefault) {
+  return <div className="p-8">{children}</div>;
+}
