--- src/app/(dashboard-routes)/dashboard/components/Menu/SubMenu.tsx
@@ -1,8 +1,0 @@
-interface SubMenuProps {
-  children: React.ReactNode;
-}
-
-function SubMenu({ children }: SubMenuProps) {
-  return <>{children}</>;
-}
-export default SubMenu;
