+++ src/app/(dashboard-routes)/dashboard/modules/page.tsx
@@ -1,4 +1,4 @@
-import { ModuleItem } from '../components/ModulesItem';
+import { ModuleItem } from '../../components/ModulesItem';
 
 export default function Modules() {
   const arrayData = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11];
